fix: size_bytes is usize not u64

+++ crates/q-types/src/legacy.rs
@@ -522,7 +522,7 @@ fn parse_old_dag_block_manual(data: &[u8]) -> Result<QBlock, bincode::Error> {
         quantum_metadata: QuantumMetadata::default(),
         transactions: vec![],
         balance_updates: vec![],
-        size_bytes: data.len() as u64,
+        size_bytes: data.len(),
     })
 }
